Removed object fit from masthead and moved it to query's
I had to move the object fit; contain code as it made the web-view masthead not fit the screen.

Instead, I added this to one of the media querys, and adjusted the min-width amount so it took affect across the min and max width of the website where the masthead was not filling the whole screen.