Changed web page file names
I changed the page file names from the default “page-one.htm” etc to match the names of each page. This is a good practise to get into as it offers clearer navigation during the development process on VS Code. I’ve left the index page alone as it is common practise to do so.